fix: some refactoring and bug crushing :)
