+++ src/components/Messages.jsx
@@ -1,70 +1,20 @@
 import { useRouter } from "next/router";
 import { useState, useEffect, useContext } from "react";
-import {
-  Box,
-  Card,
-  Fab,
-  Zoom,
-  IconButton,
-  Pagination,
-  useTheme,
-} from "@mui/material";
-import { amber } from "@mui/material/colors";
-import { getCookie, checkCookie } from "../lib/cookieMonster";
+import { Box, Pagination, useTheme } from "@mui/material";
+import { checkCookie } from "../lib/cookieMonster";
 import { refreshToken } from "../lib/browserMonster";
-import PersonIcon from "@mui/icons-material/Person";
-import EventIcon from "@mui/icons-material/Event";
-import StarBorderIcon from "@mui/icons-material/StarBorder";
-import StarIcon from "@mui/icons-material/Star";
-import KeyboardArrowUpIcon from "@mui/icons-material/KeyboardArrowUp";
 import MessageSkeleton from "./MessageSkeleton";
 import { navPrefsContext, notifContext } from "../pages/_app";
-
-const colors = {
-  Information: "#4caf50",
-  Important: "#ff9800",
-  Urgent: "#f44336",
-};
-
-const getTimePassed = (date) => {
-  // return one-seven days ago else the date
-  const now = new Date();
-  const then = new Date(date);
-  const diff = now - then;
-  const days = Math.floor(diff / (1000 * 60 * 60 * 24));
-
-  if (days < 1) {
-    return "Today";
-  } else if (days < 2) {
-    return "Yesterday";
-  } else if (days < 7) {
-    return `${days} days ago`;
-  } else {
-    return date;
-  }
-};
+import ScrollToTopFab from "./ScrollToTopFab";
+import MessageCard from "./MessageCard";
+import { getApiURL } from "../lib/util";
 
 const Messages = ({ boardID }) => {
+  // essentials
   const router = useRouter();
-
-  // Control visibility floating action button to go back up
-  const [isVisible, setIsVisible] = useState(false);
-  useEffect(() => {
-    document.querySelector(".contentframe").addEventListener("scroll", (e) => {
-      listenToScroll(e);
-    });
-    return () =>
-      document
-        .querySelector(".contentframe")
-        .removeEventListener("scroll", (e) => {
-          listenToScroll(e);
-        });
-  }, []);
-  const listenToScroll = (e) => {
-    if (e.target.scrollTop > 200) {
-      setIsVisible(true);
-    } else setIsVisible(false);
-  };
+  const theme = useTheme();
+  const ctx = useContext(navPrefsContext);
+  const notif = useContext(notifContext);
 
   // Control message display
   const [messages, setMessages] = useState("");
@@ -75,7 +25,6 @@ const Messages = ({ boardID }) => {
   const { type } = router.query;
 
   const fetchMessages = async (type, setM = true) => {
-    setLoading(true);
     // immediately refresh token if it has expired already
     // saves ~2s because iemb is slow...
     if (
@@ -114,11 +63,7 @@ const Messages = ({ boardID }) => {
       default:
         endpoint = "getBoard";
     }
-    const url = `https://iemb-backend.azurewebsites.net/api/${endpoint}?authToken=${encodeURI(
-      getCookie("auth_token")
-    )}&veriToken=${encodeURI(getCookie("veri_token"))}&sessionID=${encodeURI(
-      getCookie("sess_id")
-    )}&boardID=${boardID}${extraArgs}`;
+    const url = `${getApiURL(endpoint)}&boardID=${boardID}${extraArgs}`;
     const response = await fetch(url).catch((err) => {
       return notif.open("An error occured while fetching messages");
     });
@@ -139,16 +84,17 @@ const Messages = ({ boardID }) => {
 
     const data = await response.json();
 
-    // add data to localStorage
+    // add messages to localStorage
     localStorage.setItem(`${boardID}+${type}`, JSON.stringify(data.messages));
+    // add user's name to localStorage
     if (data.name) {
       localStorage.setItem("name", data.name);
     }
+    // pagination data for archived and starred boards
     if (type === "archived" || type === "starred") {
       setPage(data.currentPage);
       setTotalPages(data.totalPages);
     }
-
     if (setM) {
       setMessages(data.messages);
       notif.open("Messages fetched");
@@ -156,89 +102,36 @@ const Messages = ({ boardID }) => {
     if (type === "starred") {
       setFav(data.messages);
     }
-    setLoading(false);
   };
+  const getSortedMessages = (messages) => {
+    messages?.sort((a, b) => (a.pid < b.pid ? 1 : -1));
+    // sort by pid
+    // Fun fact: because iemb doesn't do this their messages are sorted correctly by date but not by time
 
-  const updateStarredStatus = async (pid, bid, status) => {
-    // immediately refresh token if it has expired already
-    // saves ~2s because iemb is slow...
-    if (
-      !checkCookie("auth_token") ||
-      !checkCookie("sess_id") ||
-      !checkCookie("veri_token")
-    ) {
-      return await refreshToken(
-        async () => updateStarredStatus(pid, bid, status),
-        notif.open,
-        router
+    messages = messages?.filter(
+      (message) =>
+        ctx.navPrefs?.messagePrefs.indexOf(message.urgency) !== -1 ||
+        // remove messages not of desired urgency
+        message.read === null
+    );
+
+    if (ctx.navPrefs?.messagePrefs.indexOf("Read") === -1) {
+      // remove messages read already
+      messages = messages?.filter((message) => {
+        if (message.read === null) {
+          return true;
+        } else return !message.read;
+      });
+    }
+
+    if (ctx.navPrefs?.messagePrefs.indexOf("ECG") === -1) {
+      // remove messages pertaining to ECG
+      messages = messages?.filter(
+        (message) => message.subject.indexOf("ECG") === -1
       );
     }
 
-    const url = `https://iemb-backend.azurewebsites.net/api/star?authToken=${encodeURIComponent(
-      getCookie("auth_token")
-    )}&veriToken=${encodeURIComponent(
-      getCookie("veri_token")
-    )}&sessionID=${encodeURIComponent(
-      getCookie("sess_id")
-    )}&bid=${bid}&pid=${pid}&status=${status}`;
-    const response = await fetch(url).catch((err) => {
-      return notif.open(
-        "An error occured while updating starred status for post " + pid,
-        "error"
-      );
-    });
-
-    switch (response.status) {
-      case 401:
-        return await refreshToken(
-          async () => updateStarredStatus(pid, bid, status),
-          notif.open,
-          router
-        );
-      case 200:
-        if (status) {
-          // star smth previously unstarred
-          let newStarredMsg = messages.find(
-            (elem) => elem.pid.toString() === pid.toString()
-          );
-          setFav([newStarredMsg, ...fav]);
-          localStorage.setItem(`${bid}+starred`, JSON.stringify(fav));
-        } else {
-          // unstar smth previously starred
-          let newStarredMsgs = fav.filter(
-            (elem) => elem.pid.toString() !== pid.toString()
-          );
-
-          setFav(newStarredMsgs);
-          localStorage.setItem(`${bid}+starred`, JSON.stringify(fav));
-        }
-        return notif.open(
-          `Successfully ${status ? "starred" : "unstarred"} post ${pid}`,
-          "success"
-        );
-      default:
-        // also handles response code 500
-        return notif.open(
-          "An error occured while updating starred status for post " + pid,
-          "error"
-        );
-    }
-  };
-
-  const updateFavStatus = async (pid) => {
-    try {
-      let curr = fav.find((elem) => elem.pid.toString() === pid.toString());
-      if (curr) {
-        await updateStarredStatus(pid, boardID, false);
-      } else {
-        await updateStarredStatus(pid, boardID, true);
-      }
-    } catch (e) {
-      notif.open(
-        "An error occured while updating starred status for post " + pid,
-        "error"
-      );
-    }
+    return messages;
   };
 
   // initiates fetching messages
@@ -267,24 +160,19 @@ const Messages = ({ boardID }) => {
     }
   }, [type]);
 
-  // fetches messages again if the user navigates to a different board
+  // fetches messages again if the user navigates to a different page
   // for archived and starred board only
   useEffect(() => {
     if (router.query.type === "archived" || router.query.type === "starred") {
       setMessages("");
-      setLoading(true);
       fetchMessages(router.query.type, page);
     }
   }, [page]);
 
-  // toggles the loading spinner
+  // toggles loading state
   useEffect(() => {
-    if (!!messages) setLoading(false);
+    setLoading(!messages);
   }, [messages]);
-
-  const theme = useTheme();
-  const ctx = useContext(navPrefsContext);
-  const notif = useContext(notifContext);
 
   return (
     <Box
@@ -307,102 +195,16 @@ const Messages = ({ boardID }) => {
             </h2>
           )}
           {!!messages &&
-            messages
-              ?.sort((a, b) => (a.pid < b.pid ? 1 : -1))
-              // sort by pid
-              // Fun fact: because iemb doesn't do this their messages are sorted correctly by date but not by time
-              ?.filter(
-                (message) =>
-                  ctx.navPrefs?.messagePrefs.indexOf(message.urgency) !== -1 ||
-                  message.read === null
-              )
-              ?.filter((message) => {
-                if (message.read === null) {
-                  return true;
-                }
-                if (
-                  ctx.navPrefs?.messagePrefs.indexOf("Read") === -1 &&
-                  message.read
-                ) {
-                  // don't include the read messages
-                  return false;
-                } else {
-                  return true;
-                }
-              })
-              ?.map((message) => (
-                <div className="messages__item" key={message.pid}>
-                  <Card
-                    variant="outlined"
-                    className={`messages__item__content ${
-                      message.read ? "read-msg" : "unread-msg"
-                    }`}
-                    sx={{
-                      borderLeft: `5px solid ${
-                        colors[message.urgency]
-                          ? colors[message.urgency]
-                          : "#ce9eff"
-                      }`,
-                      backgroundColor: theme.palette.background.paper,
-                      "&:hover": {
-                        backgroundColor: theme.palette.action.hover,
-                      },
-                      transition: "all 0.2s ease-in-out",
-                      cursor: "pointer",
-                    }}
-                  >
-                    <a
-                      href={`/post?boardID=${boardID}&pid=${message.pid}&type=${type}`}
-                    >
-                      <h2 className="messages__item__content__subject">
-                        {message.subject}
-                      </h2>
-                    </a>
-                    <div className="messages__item__content__info-wrapper">
-                      <div className="messages__item__content__info-item">
-                        <span className="messages__item__content__info-item-icon">
-                          <PersonIcon fontSize="small" />
-                        </span>
-                        <span className="messages__item__content__info-item-field">
-                          {message.username ? message.username : message.sender}
-                        </span>
-                      </div>
-                      <div className="messages__item__content__info-item">
-                        <span className="messages__item__content__info-item-icon">
-                          <EventIcon fontSize="small" />
-                        </span>
-                        <span className="messages__item__content__info-item-field">
-                          {getTimePassed(message.date)}
-                        </span>
-                      </div>
-                      <div className="messages__item__content__info-item">
-                        <span className="messages__item__content__info-item-field">
-                          <IconButton
-                            aria-label="star"
-                            size="small"
-                            onClick={() => {
-                              updateFavStatus(message.pid);
-                            }}
-                          >
-                            {fav?.find(
-                              (elem) =>
-                                elem.pid.toString() === message.pid.toString()
-                            ) ? (
-                              <StarIcon
-                                fontSize="small"
-                                sx={{ color: amber[500] }}
-                              />
-                            ) : (
-                              <StarBorderIcon fontSize="small" />
-                            )}
-                          </IconButton>
-                        </span>
-                      </div>
-                    </div>
-                  </Card>
-                </div>
-              ))}
-
+            getSortedMessages(messages)?.map((message) => (
+              <MessageCard
+                key={message.pid}
+                message={message}
+                fav={fav}
+                setFav={setFav}
+                messages={messages}
+                boardtype={type}
+              />
+            ))}
           {!!messages &&
             (router.query.type === "archived" ||
               router.query.type === "starred") && (
@@ -424,34 +226,7 @@ const Messages = ({ boardID }) => {
                 />
               </div>
             )}
-          <Zoom in={isVisible} timeout={300}>
-            <Fab
-              size="medium"
-              color="secondary"
-              aria-label="back-to-top"
-              sx={{
-                margin: 0,
-                top: "auto",
-                right: "2rem",
-                bottom: "3.2rem",
-                left: "auto",
-                position: "fixed",
-                backgroundColor: "#ce9eff",
-                "&:hover": {
-                  backgroundColor: "#b46bff",
-                },
-              }}
-              onClick={() => {
-                document.querySelector(".contentframe").scrollTo({
-                  top: 0,
-                  left: 0,
-                  behavior: "smooth",
-                });
-              }}
-            >
-              <KeyboardArrowUpIcon />
-            </Fab>
-          </Zoom>
+          <ScrollToTopFab />
         </Box>
       )}
     </Box>
