+++ src/lib/cookieMonster.js
