+++ src/styles/components/_Login.scss
@@ -119,7 +119,7 @@
   }
   &__content-bg {
     box-shadow: -5px 0 10px rgba(0, 0, 0, 0.6);
-    border-radius: 5px;
+    border-radius: 5px 0 0 5px;
     background-color: rgba(0, 0, 0, 0.6);
 
     width: 80%;
